Añadido feedback al:  - Eliminar un juego.  - Eliminar un jugador.  - Editar un juego.  - Editar un jugador.  - Crear un juego.  - Crear un jugador.  - Seguir a un jugador.  - Dejar de seguir a un jugador.  - Establecer un juego como favorito.  - Quitar un juego de favoritos.  - En caso de que alguna lista esté vacía (sobre todo al buscar).
Faltante:
 - Terminar de ajustar el feedback al crear una noticia.
 - Comprobar los campos al crear un juego, jugador o noticia.
 - Dudas puntuales.

+++ interfaz/src/main/resources/static/js/gameTable.js
@@ -22,30 +22,59 @@
 
 function updateFavourite(checkIdName, gameId)
 {
+	var options = {
+			settings: {
+				duration: 2000
+			}
+		};
+	
 	 if($('#' + checkIdName + gameId).is(':checked'))
-         $.post('http://localhost:8083/players/1/games/?id=' + gameId);
-     else
+	 {
+         $.post('http://localhost:8083/players/1/games/?id=' + gameId, function() {
+        	 iqwerty.toast.Toast('Juego con ID ' + gameId +  ' añadido a favoritos.', options);
+         });
+     }else
+     {
     	 $.ajax({
     		    url: 'http://localhost:8083/players/1/games/' + gameId,
     		    method: 'DELETE',
     		    success: function(result) {
-    		        // Do something with the result
+    		    	iqwerty.toast.Toast('Juego con ID ' + gameId +  ' eliminado de favoritos.', options);
     		    }
     		});
+     }
 };
 
 function updateFollow(checkIdName, playerId)
 {
+	var options = {
+			settings: {
+				duration: 2000
+			}
+		};
+	
+	if(playerId == 1)
+	{
+		document.getElementById(checkIdName + playerId).checked = false;
+		iqwerty.toast.Toast('No puedes seguirte a ti mismo.', options);
+		return;
+	}
+	
 	 if($('#' + checkIdName + playerId).is(':checked'))
-         $.post('http://localhost:8084/players/1/followers/?id=' + playerId);
-     else
+	 {
+         $.post('http://localhost:8084/players/1/followers/?id=' + playerId, function() {
+        	 iqwerty.toast.Toast('Jugador con ID ' + playerId +  ' seguido.', options);
+         });
+	 }else
+	 {
     	 $.ajax({
     		    url: 'http://localhost:8084/players/1/followers/' + playerId,
     		    method: 'DELETE',
     		    success: function(result) {
-    		        // Do something with the result
+    		    	iqwerty.toast.Toast('Jugador con ID ' + playerId +  ' dejado de seguir.', options);
     		    }
     		});
+	 }
 };
 
 function searchPlayer(evt)
@@ -53,6 +82,8 @@ function searchPlayer(evt)
 	evt.preventDefault();
 	var prefix = $('input[name=searchRadio]:checked').val();
 	var searchInput = $("#searchInput").val();
+	
+	if(searchInput.localeCompare("") == 0) searchInput = "null";
 	
 	if(prefix == "nameRadio")
 	{
@@ -77,6 +108,8 @@ function searchGame(evt)
 	evt.preventDefault();
 	var prefix = $('input[name=searchRadio]:checked').val();
 	var searchInput = $("#searchInput").val();
+	
+	if(searchInput.localeCompare("") == 0) searchInput = "null";
 	
 	if(prefix == "nameRadio")
 	{
